+++ PopupBridge/src/main/java/com/braintreepayments/api/PopupBridgeNavigationListener.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.popupbridge;
+package com.braintreepayments.api;
 
 public interface PopupBridgeNavigationListener {
     void onUrlOpened(String url);